+++ lizmap/table_manager/base.py
